patricia: fix bounds error in ascii2prefix
When parsing network masks, we forget to update the length of
the IP address before scanning it for ':' to detect if an
address is IPv6 or not.  This caused memchr to falsely detect
':' past the end of string (but not the allocated array) if the
stack had a dirty ':' leftover.

Falsely detecting ':' leads to AF_INET6 being passed to
inet_pton() on some AF_INET (IPv4) addresses, causing
a failure.

Fortunately this bug could not cause a segfault or bus error
(and thus went undetected by valgrind).   The code was always
accessing memory within the stack allocated in the "save" array
of INET6_ADDRSTRLEN bytes.